+++ frontend/src/components/Deployer.tsx
@@ -1,0 +1,124 @@
+import React, { useEffect, useState } from 'react';
+import Urbit from '@urbit/http-api';
+
+interface DeskMetadata {
+  desk: string;
+}
+
+const api = new Urbit('', '', window.desk);
+api.ship = window.ship;
+
+export function Deployer() {
+  // @ts-ignore
+  const [desks, setDesks] = useState<Array<string>>();
+
+  useEffect(() => {
+    async function init() {
+
+      api.subscribe({
+        app: "vita-deploy",
+        path: "/frontend",
+        event: handleSub,
+        quit: () => alert("lost connection to your urbit. please refresh"),
+        err: (e) => console.log("error", e),
+      }).then((subscriptionId) => {
+        //
+        window.addEventListener("beforeunload", () => {
+          api.unsubscribe(subscriptionId);
+          api.delete();
+        });
+      });
+    }
+
+    init();
+  }, []);
+
+  function handleSub(data: any) {
+    console.log('got update', data)
+    const desks = 'desks'
+    if (data[desks] !== undefined) {
+      const newDesks: Array<string> = data[desks]
+      setDesks(newDesks);
+    }
+  }
+
+  const invalidDeskNames = ['base', 'landscape', 'groups', 'talk', 'realm', 'vita', 'lemur', 'pals', 'radio', 'rumors', 'portal']
+  function newDesk(deskName: string) {
+    if (invalidDeskNames.includes(deskName)) {
+      return;
+    }
+    api.poke({
+      app: "vita-deploy",
+      mark: "vita-deploy-action",
+      json: {
+        'create-app': deskName
+      },
+    });
+
+  }
+
+  return (
+    <div>
+      <a href="/vita"> &lt;- app metrics </a>
+      <h1>vita app deployer</h1>
+      <div>
+        <div>
+
+          <div>
+
+            <input type="text" id="new-desk-name"
+              placeholder="unique-desk-name"
+
+            />
+            <button onClick={() => {
+              const deskName = (document.getElementById('new-desk-name') as HTMLInputElement).value;
+
+              let element = document.getElementById('new-desk-name') as HTMLInputElement;
+              if (element) {
+                element.value = '';
+              }
+
+              newDesk(deskName)
+
+            }}
+            >
+              create app
+            </button>
+          </div>
+        </div>
+
+
+        <div>
+          <ul>
+            {/* created desks */}
+            {desks?.map((desk: string) => (
+              <li key={desk} >
+                <div
+                  style={{
+                    display: 'flex',
+                    flexDirection: 'row',
+                    gap: '1rem',
+                    padding: '0.5rem',
+                    margin: '0.5rem 0'
+                  }}
+                >
+                  <div>
+                    {'%'}{desk}
+                  </div>
+                  <div>
+                    <a href={`/apps/vita/config/${desk}`} >
+                      config
+                    </a>
+                  </div>
+                </div>
+              </li>
+            ))}
+          </ul>
+        </div>
+
+
+      </div>
+
+    </div >
+  );
+}
