+++ frontend/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { App } from './app';
+import App from './app'; // Fixed import statement
 import './index.css';
 
 ReactDOM.render(
